Add files via upload
Write a program that receives a 2-dimensional n∗n array using only pointers, then receives a number and if there is a number in the coordinate array, it prints the message not found.

In this question, you are asked to define the array with pointers! (don't use [] to define)

Investigate how to initialize a two-dimensional array using a pointer!

If there is more than one number equal to the requested number, print only the coordinates of the first higher and then the leftmost number.